feat(dashboard): add real-time user activity streak calculation & UI updates
Implemented the core backend logic to calculate and store user daily login streaks using native PostgreSQL gaps-and-islands queries. The frontend dashboard now dynamically updates the streak and activity status based on real API data. Also resolved a critical JWT ClassCastException and refactored user point to score.

+++ backend/src/main/java/com/swp391/coding_platform/repository/course/EnrollmentRepository.java
@@ -16,9 +16,9 @@ import java.util.Set;
 @Repository
 public interface EnrollmentRepository extends JpaRepository<EnrollmentEntity, Long> {
     Boolean existsByUserIdAndCourseIdAndStatusIn(Long userId, Long courseId, Collection<EnrollmentStatus> statuses);
-    Long countByUserId(Long userId);
+    Long countByUserId(Integer userId);
 
-    Long countByUserIdAndStatus(Long userId, EnrollmentStatus status);
+    Long countByUserIdAndStatus(Integer userId, EnrollmentStatus status);
 
     @Query("SELECT e.course.id FROM EnrollmentEntity e " +
             "WHERE e.user.id = :userId " +
